# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello Swagger App
# during dev, should point to your local machine

securityDefinitions:
  Bearer:
    description: OpenStack Keystone Authentication
    type: apiKey
    name: Authorization
    in: header
  apiKey:
    description: OpenStack Keystone Authentication
    type: apiKey
    name: Authorization
    in: header    
  basicAuth:
    type: basic


security:
  - basicAuth: []
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
  - application/x-www-form-urlencoded
  - image/gif
  - image/png
  - image/jpg
  - image/jpeg  

# format of the responses to the client (Accepts)
produces:
  - application/json
  - multipart/form-data 
  - application/x-www-form-urlencoded
  - image/gif
  - image/png
  - image/jpg
  - image/jpeg
paths:
  /hello:
    # binds a127 app logic to a route

    x-swagger-router-controller: hello
    get:
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
          default: jaeguk
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /test:
    # binds a127 app logic to a route

    x-swagger-router-controller: hello
    get: 
      description: Returns 'Hello' to the caller
      security:
        - apiKey: []      
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
          default: jaeguk
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            

  /upload:
    x-swagger-router-controller: hello
    post:
      summary: Uploads a file.
      consumes:
        - multipart/form-data
      operationId: createImage
      parameters:
        - in: formData
          name: account_copy
          description: The file to upload.
          type: file

        - in: formData
          name: biz_regist_no_copy
          type: file
          description: The file to upload.          
      responses:
        201: 
          description: SuccessFul request
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /uploadToDB:
    x-swagger-router-controller: hello
    post:
      summary: Uploads a file.
      consumes:
        - multipart/form-data
      operationId: uploadToDB
      parameters:
        - in: formData
          name: account
          description: The file to upload.
          type: file

        - in: formData
          name: biz_regist_no
          type: file
          description: The file to upload.          
      responses:
        201: 
          description: SuccessFul request
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            

  /downloadFromFile:
    x-swagger-router-controller: hello   
    get:
        summary: download a file.
        consumes:
          - "application/json"
        produces:
          - image/gif
          - image/png
          - image/jpg
          - image/jpeg
        operationId: download
        parameters:
          - in: query
            name: filename
            description: The file name.
            type: string       
        responses:
          209: 
            description: SuccessFul request
            schema:
              type: file

          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"      
 
  /downloadFromDB:
    x-swagger-router-controller: hello   
    get:
        summary: download a file.
        consumes:
          - "application/json"
        produces:
          - image/gif
          - image/png
          - image/jpg
          - image/jpeg
        operationId: downloadFromDB
        parameters:
          - in: query
            name: filename
            description: The file name.
            type: string       
        responses:
          209: 
            description: SuccessFul request
            schema:
              type: file

          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"                
   
  /token:
    x-swagger-router-controller: hello
    post:   
      summary: token
      x-security-scopes:
        - admin
        - user    
      consumes:
        - application/json
      produces:
        - application/json
      operationId: token
      parameters:
        - name: user_type
          in: body
          required: true   
          description: User's type.
          schema:
            properties:
              user_type:
                type: string
              username:
                type: string
              password:
                type: string

      responses:
        201: 
          description: SuccessFul request
          schema:
            properties:
              token:
                type: string      

  /login:
    x-swagger-router-controller: hello
    post:   
      summary: login
      security:
        - basicAuth: []
      x-security-scopes:
        - admin
        - user    
      consumes:
        - application/json
      produces:
        - application/json
      operationId: auth
      parameters:
        - name: user_type
          in: body
          required: true   
          description: User's type.
          schema:
            properties:
              user_type:
                type: string
              username:
                type: string
              password:
                type: string
         
      responses:
        201: 
          description: SuccessFul request
          schema:
            $ref: "#/definitions/SuccessResponse"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string

  SuccessResponse:
    properties:
      id:
        type: string
      extension:
        type: string
      size:
        type: integer
      type:
        type: string

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string